Move context tests into their own file
